Move process_bundles to domain_block_processor This commit also delete system_bundle_processor/core_bundle_processor as they are not used anymore.
