+++ domains/client/domain-executor/src/core_domain_worker.rs
@@ -14,9 +14,10 @@
 // You should have received a copy of the GNU General Public License
 // along with Polkadot.  If not, see <http://www.gnu.org/licenses/>.
 
-use crate::core_bundle_processor::CoreBundleProcessor;
 use crate::core_bundle_producer::CoreBundleProducer;
+use crate::domain_block_processor::DomainBlockProcessor;
 use crate::domain_worker::{handle_block_import_notifications, handle_slot_notifications};
+use crate::parent_chain::CoreDomainParentChain;
 use crate::utils::{BlockInfo, ExecutorSlotInfo};
 use crate::TransactionFor;
 use domain_runtime_primitives::{AccountId, DomainCoreApi, DomainExtrinsicApi};
@@ -37,6 +38,7 @@ use system_runtime_primitives::SystemDomainApi;
 use tracing::Instrument;
 
 #[allow(clippy::too_many_arguments)]
+#[allow(clippy::type_complexity)]
 pub(super) async fn start_worker<
     Block,
     SBlock,
@@ -53,13 +55,15 @@ pub(super) async fn start_worker<
     primary_chain_client: Arc<PClient>,
     client: Arc<Client>,
     bundle_producer: CoreBundleProducer<Block, SBlock, PBlock, Client, SClient, TransactionPool>,
-    bundle_processor: CoreBundleProcessor<
+    domain_block_processor: DomainBlockProcessor<
         Block,
-        SBlock,
         PBlock,
+        SBlock,
+        SBlock,
         Client,
-        SClient,
         PClient,
+        SClient,
+        CoreDomainParentChain<SClient, SBlock, PBlock>,
         Backend,
         E,
     >,
@@ -112,7 +116,7 @@ pub(super) async fn start_worker<
                 let span = span.clone();
 
                 move |primary_info| {
-                    bundle_processor
+                    domain_block_processor
                         .clone()
                         .process_bundles(primary_info)
                         .instrument(span.clone())
