+++ domains/client/domain-executor/src/system_executor.rs
@@ -1,6 +1,6 @@
 use crate::domain_block_processor::DomainBlockProcessor;
 use crate::fraud_proof::FraudProofGenerator;
-use crate::system_bundle_processor::SystemBundleProcessor;
+use crate::parent_chain::SystemDomainParentChain;
 use crate::system_bundle_producer::SystemBundleProducer;
 use crate::{active_leaves, EssentialExecutorParams, TransactionFor};
 use domain_runtime_primitives::{AccountId, DomainCoreApi, DomainExtrinsicApi};
@@ -20,6 +20,7 @@ use subspace_core_primitives::Blake2b256Hash;
 use system_runtime_primitives::SystemDomainApi;
 
 /// System domain executor.
+#[allow(clippy::type_complexity)]
 pub struct Executor<Block, PBlock, Client, PClient, TransactionPool, Backend, E>
 where
     PBlock: BlockT,
@@ -30,7 +31,18 @@ where
     transaction_pool: Arc<TransactionPool>,
     backend: Arc<Backend>,
     fraud_proof_generator: FraudProofGenerator<Block, PBlock, Client, Backend, E>,
-    bundle_processor: SystemBundleProcessor<Block, PBlock, Client, PClient, Backend, E>,
+    domain_block_processor: DomainBlockProcessor<
+        Block,
+        PBlock,
+        Block,
+        PBlock,
+        Client,
+        PClient,
+        Client,
+        SystemDomainParentChain<PClient, Block, PBlock>,
+        Backend,
+        E,
+    >,
 }
 
 impl<Block, PBlock, Client, PClient, TransactionPool, Backend, E> Clone
@@ -46,7 +58,7 @@ where
             transaction_pool: self.transaction_pool.clone(),
             backend: self.backend.clone(),
             fraud_proof_generator: self.fraud_proof_generator.clone(),
-            bundle_processor: self.bundle_processor.clone(),
+            domain_block_processor: self.domain_block_processor.clone(),
         }
     }
 }
@@ -129,22 +141,17 @@ where
             params.code_executor,
         );
 
+        let parent_chain = SystemDomainParentChain::new(params.primary_chain_client.clone());
+
         let domain_block_processor = DomainBlockProcessor::new(
             DomainId::SYSTEM,
             params.client.clone(),
             params.primary_chain_client.clone(),
             params.client.clone(),
+            parent_chain,
             params.primary_network,
             params.backend.clone(),
             fraud_proof_generator.clone(),
-        );
-
-        let bundle_processor = SystemBundleProcessor::new(
-            params.primary_chain_client.clone(),
-            params.client.clone(),
-            params.backend.clone(),
-            params.keystore,
-            domain_block_processor,
         );
 
         spawn_essential.spawn_essential_blocking(
@@ -154,7 +161,7 @@ where
                 params.primary_chain_client.clone(),
                 params.client.clone(),
                 bundle_producer,
-                bundle_processor.clone(),
+                domain_block_processor.clone(),
                 params.imported_block_notification_stream,
                 params.new_slot_notification_stream,
                 active_leaves,
@@ -170,7 +177,7 @@ where
             transaction_pool: params.transaction_pool,
             backend: params.backend,
             fraud_proof_generator,
-            bundle_processor,
+            domain_block_processor,
         })
     }
 
@@ -186,7 +193,11 @@ where
         self,
         primary_info: (PBlock::Hash, NumberFor<PBlock>, ForkChoiceStrategy),
     ) {
-        if let Err(err) = self.bundle_processor.process_bundles(primary_info).await {
+        if let Err(err) = self
+            .domain_block_processor
+            .process_bundles(primary_info)
+            .await
+        {
             tracing::error!(?primary_info, ?err, "Error at processing bundles.");
         }
     }
